Rename 13._RomanToInteger_Solution1 to 13RomanToInteger_Solution1
Roman numerals are represented by seven different symbols: I, V, X, L, C, D and M.

Symbol       Value
I             1
V             5
X             10
L             50
C             100
D             500
M             1000

For example, two is written as II in Roman numeral, just two one's added together. Twelve is written as, XII, which is simply X + II. The number twenty seven is written as XXVII, which is XX + V + II.

Roman numerals are usually written largest to smallest from left to right. However, the numeral for four is not IIII. Instead, the number four is written as IV. Because the one is before the five we subtract it making four. The same principle applies to the number nine, which is written as IX. There are six instances where subtraction is used:

I can be placed before V (5) and X (10) to make 4 and 9. 
X can be placed before L (50) and C (100) to make 40 and 90. 
C can be placed before D (500) and M (1000) to make 400 and 900.

solution1是一个比较智障的解法，创建hash表可以简化代码量
solution1思路：
除三种罗马数字需要讨论是否在大数的左边表示相减之外，其他都是加上对应的阿拉伯数字的值。
再来处理3中特殊情况，分别判断在I、X、C右边出现的罗马符号，若对应的是V或X，L或C，D或M则int值是减去相应的值。